requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dateutil>=2.9.0
